extract ComplianceTest weaver based framework
